+huang1111filepvder

+++ Aquc.AquaUpdater/Aquc.AquaUpdater/Structure.cs
@@ -67,8 +67,10 @@ public struct UpdatePackage
         updateSubscription.lastCheckUpdateTime = DateTime.Now;
         updateSubscription.currentlyVersion = updateMessage.packageVersion;
         Launch.UpdateLaunchConfig();
-        zipPath.Delete();
-        extraceZipDirectory.Delete(true);
+        if (extraceZipDirectory.GetFiles(".donotremovezip").Length!=0)
+            zipPath.Delete();
+        if (extraceZipDirectory.GetFiles(".donotremoveextracezip").Length != 0)
+            extraceZipDirectory.Delete(true);
     }
     private void CopyDirectory(DirectoryInfo directory, DirectoryInfo dest)
     {
